feat: support for multiple crontab entries (#10)
<!--
To streamline the process of creating and reviewing pull requests, we
have created a template.
It is not required to follow this template perfectly, but we encourage
you to think about each header.
Before you publish a pull request think about the definition of done for
the issue you are trying to resolve.
-->

<!--
Provide the key for the Jira issue this PR resolves.
-->

Closes PM-

## Description
<!--
Provide a concise description on the feature/fix your pull request
implements.
-->

## Changes
<!--
Provide a list of important changes.

e.g.
* Added tests
* Renamed variable
* Added topic
-->

## Testing

<!-- Provide additional information necessary for testing this PR. This
includes details like branches in other repos, launch arguments or gait
directories. In the case of a bug fix, provide the steps to reproduce
the bug.-->

<!-- Reviews are automatically requested after the pull request has been
created. Only request for a review if you want a specific person to
review your changes.-->